wpa_suplicant: fix for interface not found
Problem: Pressing "On" button from the Settings application the HAL is
loading the wl12xx.ko module but in the same time WifiStateMachine receives
the command to start the wpa_supplicant through HAL. It has been observed
that the wpa_supplicant is trying to get the wlan0 interface but for few
seconds the kernel hasn't exposed it yet.

So will add a retry for it.

Signed-off-by: Nicusor Huhulea <nicusor_huhulea@mentor.com>